chore(java): ProtobufDispatcher should trigger log warning at most once (#2353)
<!--
**Thanks for contributing to Fory.**

**If this is your first time opening a PR on fory, you can refer to
[CONTRIBUTING.md](https://github.com/apache/fory/blob/main/CONTRIBUTING.md).**

Contribution Checklist

- The **Apache Fory (incubating)** community has restrictions on the
naming of pr titles. You can also find instructions in
[CONTRIBUTING.md](https://github.com/apache/fory/blob/main/CONTRIBUTING.md).

- Fory has a strong focus on performance. If the PR you submit will have
an impact on performance, please benchmark it first and provide the
benchmark result here.
-->

## What does this PR do?

This PR addresses these 2 issues:
- It's possible the `ProtobufSerializer not loaded...` log warning to
fire even if you aren't trying to serialise a Protobuf class.
- The warning is fired every time
`ProtobufSerializer.getSerializerClass()` is called which can result in
the log line to be constantly fired and displayed in logs.

With these changes, the log warning can only be fired once and only the
first time you try to serialise a Protobuf class.

## Related issues

This has been discussed in
https://github.com/apache/fory/discussions/2349

## Does this PR introduce any user-facing change?
- [ ] Does this PR introduce any public API change?
- [ ] Does this PR introduce any binary protocol compatibility change?

## Benchmark

I believe it should have an insubstantial benchmark effect.