perf(Go/python): Type system compatibility upgrade (#2339)
<!--
**Thanks for contributing to Fory.**

**If this is your first time opening a PR on fory, you can refer to
[CONTRIBUTING.md](https://github.com/apache/fory/blob/main/CONTRIBUTING.md).**

Contribution Checklist

- The **Apache Fory (incubating)** community has restrictions on the
naming of pr titles. You can also find instructions in
[CONTRIBUTING.md](https://github.com/apache/fory/blob/main/CONTRIBUTING.md).

- Fory has a strong focus on performance. If the PR you submit will have
an impact on performance, please benchmark it first and provide the
benchmark result here.
-->

## What does this PR do?


In this PR, as the mentor, I completed the type system upgrade for
furygo together with @lzaeh. The key tasks include:

1、Setting up the basic Go framework and defining upgrade objectives
@pandalee99
2、Implementing basic named structs and identifying issues to resolve
@pandalee99
3、Carrying out cross-language transformation for the MAP serializer
@lzaeh
4、Conducting cross-language transformation for the Slice serializer
@lzaeh
5、Detecting and fixing errors in structure hash calculation @lzaeh
6、Solving cross-language serialization issues for structures @lzaeh
7、Resolving type system registration problems @lzaeh

<!-- Describe the purpose of this PR. -->

## Related issues
#2014 #2200 #2201 #2202 #2268

## Relate PR

#2198 #2230 

<!--
Is there any related issue? Please attach here.

- #xxxx0
- #xxxx1
- #xxxx2
-->

## Does this PR introduce any user-facing change?

@lzaeh, could you please elaborate on the changes made in this PR and
share your insights on the type system upgrade for furygo? It would be
great if you can walk us through the key challenges you encountered, the
solutions implemented, and the impact of these changes on the project's
functionality and cross-language compatibility. Looking forward to your
detailed explanation!

<!--
If any user-facing interface changes, please [open an
issue](https://github.com/apache/fory/issues/new/choose) describing the
need to do so and update the document if necessary.
-->

- [ ] Does this PR introduce any public API change?
- [ ] Does this PR introduce any binary protocol compatibility change?

## Benchmark

<!--
When the PR has an impact on performance (if you don't know whether the
PR will have an impact on performance, you can submit the PR first, and
if it will have impact on performance, the code reviewer will explain
it), be sure to attach a benchmark data here.
-->

---------

Co-authored-by: lzaeh <ustc.amnesia@gmail.com>
Co-authored-by: lzaeh <154985570+lzaeh@users.noreply.github.com>
Co-authored-by: O O <oo@Os-MacBook-Pro.local>