docs: Improve pyfory PyPI documentation (#2498)
## What does this PR do?

The existing PyPI documentation for `pyfory` was a single line and
uninformative. This change improves the documentation by:

1. Creating a new `python/README.md` with a clear and concise
description of the project, installation instructions, and usage
examples. This file is used to generate the PyPI page description.
2. Moving the developer-focused build and test instructions to a new
`python/CONTRIBUTING.md` file.

This will provide a much better experience for Python developers who
discover `pyfory` on PyPI, similar to the documentation for other
popular libraries like `opendal` and `msgpack`.

## Related issues

Closes #2488