feat(python): type meta encoding for python (#2509)
## Why?

type forward/backward compatible serialization is critical for online
service which different service update their data schema and deploy at
different time. The schema is in an inconsistent state.

meta shared encoding can address this :
https://fory.apache.org/docs/specification/fory_xlang_serialization_spec#type-def

## What does this PR do?

Add type meta encoding for python to support type forward/backward
compatible serialization.

Things not finished in this PR:
- not null field support
- generate serializer from type meta
- meta share mode

Those feature will be implemented in follow-up PRs.

## Related issues

#1938
#2160
#2278

## Does this PR introduce any user-facing change?

<!--
If any user-facing interface changes, please [open an
issue](https://github.com/apache/fory/issues/new/choose) describing the
need to do so and update the document if necessary.

Delete section if not applicable.
-->

- [ ] Does this PR introduce any public API change?
- [ ] Does this PR introduce any binary protocol compatibility change?

## Benchmark

<!--
When the PR has an impact on performance (if you don't know whether the
PR will have an impact on performance, you can submit the PR first, and
if it will have impact on performance, the code reviewer will explain
it), be sure to attach a benchmark data here.

Delete section if not applicable.
-->