feat(Rust): Implementing Type Compatible (#2492)
## What does this PR do?
- Modify according to the latest `xlang` specification. (Implemented but
not tested with `Java`)
- Improved the implementation of compatible mode by comparing the
`field_name` and `field_type` of the sender and receiver to determine
whether to assign or skip.

## TODO
- Haven’t figured out how to serialize and deserialize `Box` yet. 
- Handling of built-in unsigned types. (I will add type_id like
`125=>u8` to implement se/deserialize for `rust to rust`)
- Special handling of Option field types when extracting the generic
structure of field types in macros.
- In deserialization, for fields that do not exist locally and need to
be skipped, automatically create a deserializer for the corresponding
type and consume the byte data.

## Related issues
- #2145


## Does this PR introduce any user-facing change?
<!--
If any user-facing interface changes, please [open an
issue](https://github.com/apache/fory/issues/new/choose) describing the
need to do so and update the document if necessary.
-->

- [ ] Does this PR introduce any public API change?
- [x] Does this PR introduce any binary protocol compatibility change?

## Benchmark

<!--
When the PR has an impact on performance (if you don't know whether the
PR will have an impact on performance, you can submit the PR first, and
if it will have impact on performance, the code reviewer will explain
it), be sure to attach a benchmark data here.
-->