feat(Go): Implement ahead of time codegen for fory-go serialization (#2553)
## What does this PR do?

- Generates serialization code for structs using the `go/codegen`
package.
- Implements `Write` and `Read` methods for struct types to support
serialization and deserialization.
- Integrates `go:generate` directive to automate the generation of the
serialization code.

**Note:** This PR only supports basic struct types and does **not** yet
handle `map` or `slice` types.

## TODO
- Extend the code generation to handle `slice` types (planned for the
next PR).
- Implement support for `map` types (planned for a future PR).

## Related issues
* #2227 
<!--
Is there any related issue? If this PR closes them you say say
fix/closes:

- #xxxx0
- #xxxx1
- Fixes #xxxx2
-->

## Does this PR introduce any user-facing change?

<!--
If any user-facing interface changes, please [open an
issue](https://github.com/apache/fory/issues/new/choose) describing the
need to do so and update the document if necessary.

Delete section if not applicable.
-->

- [ ] Does this PR introduce any public API change?
- [ ] Does this PR introduce any binary protocol compatibility change?

## Benchmark

<!--
When the PR has an impact on performance (if you don't know whether the
PR will have an impact on performance, you can submit the PR first, and
if it will have impact on performance, the code reviewer will explain
it), be sure to attach a benchmark data here.

Delete section if not applicable.
-->

---------

Co-authored-by: Shawn Yang <chaokunyang@apache.org>