feat(java): support graalvm 25 (#2652)
## Why?

Graalvm 25 needs configure `unsafeAllocate` for object creation by
`Unsafe`, which is not easy to use.

## What does this PR do?

This PR provide a unified mechanism to create objects to support
graalvm25

## Related issues

Fixes #2647
https://github.com/oracle/graal/issues/12220
## Does this PR introduce any user-facing change?

<!--
If any user-facing interface changes, please [open an
issue](https://github.com/apache/fory/issues/new/choose) describing the
need to do so and update the document if necessary.

Delete section if not applicable.
-->

- [ ] Does this PR introduce any public API change?
- [ ] Does this PR introduce any binary protocol compatibility change?

## Benchmark

<!--
When the PR has an impact on performance (if you don't know whether the
PR will have an impact on performance, you can submit the PR first, and
if it will have impact on performance, the code reviewer will explain
it), be sure to attach a benchmark data here.

Delete section if not applicable.
-->