fix(java): retain serializers in graalvm native images (#2680)
## Why?

Custom serializers are not retained during GraalVM native image builds.

## What does this PR do?

The original code checks against `initialClassInfos` and removes
serializers outside the set during native image builds. This PR removes
that as well as `initialClassInfos`.

I don't know what `initialClassInfos` is for, because it is only
involved in clearing serializers, which this PR is to remove.

Also, since the tests before weren't retaining some serializers, it
turns out the original `native-image.properties` might be incomplete.
And I added more classes to `--initialize-at-build-time` as is hinted by
the `native-image` command.

## Related issues

- Fixes #2669