fix(java): fix deserializing EXT type bug (#2700)
<!--
**Thanks for contributing to Apache Fory™.**

**If this is your first time opening a PR on fory, you can refer to
[CONTRIBUTING.md](https://github.com/apache/fory/blob/main/CONTRIBUTING.md).**

Contribution Checklist

- The **Apache Fory™** community has requirements on the naming of pr
titles. You can also find instructions in
[CONTRIBUTING.md](https://github.com/apache/fory/blob/main/CONTRIBUTING.md).

- Apache Fory™ has a strong focus on performance. If the PR you submit
will have an impact on performance, please benchmark it first and
provide the benchmark result here.
-->

## Why?

<!-- Describe the purpose of this PR. -->
When the call order is register(Class, int)→ registerSerializer(Class,
Serializer), the internal state of the framework is inconsistent:
register() creates the initial TypeInfo and assigns xtypeId
registerSerializer() to update classInfo.xtypeId and serializer
But the old map in xtypeIdToClassMap is not cleaned
Causes a Type id 26387 not registered exception to be thrown during
deserialization.

## What does this PR do?

<!-- Describe the details of this PR. -->
Before update: Remove the old xtypeId map
After update: Adding a new xtypeId map ensures idempotence registration
## Related issues
Close #2698 
<!--
Is there any related issue? If this PR closes them you say say
fix/closes:

- #xxxx0
- #xxxx1
- Fixes #xxxx2
-->

## Does this PR introduce any user-facing change?

<!--
If any user-facing interface changes, please [open an
issue](https://github.com/apache/fory/issues/new/choose) describing the
need to do so and update the document if necessary.

Delete section if not applicable.
-->

- [x] Does this PR introduce any public API change?
- [x] Does this PR introduce any binary protocol compatibility change?

## Benchmark

<!--
When the PR has an impact on performance (if you don't know whether the
PR will have an impact on performance, you can submit the PR first, and
if it will have impact on performance, the code reviewer will explain
it), be sure to attach a benchmark data here.

Delete section if not applicable.
-->