docs(rust): add detailed rust api doc and readme (#2728)
## Why?

<!-- Describe the purpose of this PR. -->

## What does this PR do?

1. update rust api doc and readme
2. re-export core API
3. fix `Vec` with `Rc/Arc` trait

## Related issues

#2723 

## Does this PR introduce any user-facing change?

<!--
If any user-facing interface changes, please [open an
issue](https://github.com/apache/fory/issues/new/choose) describing the
need to do so and update the document if necessary.

Delete section if not applicable.
-->

- [ ] Does this PR introduce any public API change?
- [ ] Does this PR introduce any binary protocol compatibility change?

## Benchmark

<!--
When the PR has an impact on performance (if you don't know whether the
PR will have an impact on performance, you can submit the PR first, and
if it will have impact on performance, the code reviewer will explain
it), be sure to attach a benchmark data here.

Delete section if not applicable.
-->