fix(java): serialize collections with only null elements in xlang mode (#2740)
## What does this PR do?
When all collection elements are null, elemClass is set to Object.class
which needs to be handled by classResolver in XLANG mode.

## Related issues
Fixes #2203
<!--
Is there any related issue? If this PR closes them you say say
fix/closes:


-->

## Does this PR introduce any user-facing change?

<!--
If any user-facing interface changes, please [open an
issue](https://github.com/apache/fory/issues/new/choose) describing the
need to do so and update the document if necessary.

Delete section if not applicable.
-->

- [ ] Does this PR introduce any public API change?
- [ ] Does this PR introduce any binary protocol compatibility change?

## Benchmark

<!--
When the PR has an impact on performance (if you don't know whether the
PR will have an impact on performance, you can submit the PR first, and
if it will have impact on performance, the code reviewer will explain
it), be sure to attach a benchmark data here.

Delete section if not applicable.
-->