chore(python): rename numeric typehint (#2808)
## Why?

<!-- Describe the purpose of this PR. -->

## What does this PR do?

rename numeric typehint for the following reasons:

  1. Pythonic: Matches Python's built-in type naming conventions
2. Industry standard: Libraries like NumPy, PyArrow use this style
(np.int8, pa.int64())
3. Cleaner type hints: value: int8 is more readable than value: Int8Type
  4. Less redundant: The context already makes it clear these are types

## Related issues

<!--
Is there any related issue? If this PR closes them you say say
fix/closes:

- #xxxx0
- #xxxx1
- Fixes #xxxx2
-->

## Does this PR introduce any user-facing change?

<!--
If any user-facing interface changes, please [open an
issue](https://github.com/apache/fory/issues/new/choose) describing the
need to do so and update the document if necessary.

Delete section if not applicable.
-->

- [ ] Does this PR introduce any public API change?
- [ ] Does this PR introduce any binary protocol compatibility change?

## Benchmark

<!--
When the PR has an impact on performance (if you don't know whether the
PR will have an impact on performance, you can submit the PR first, and
if it will have impact on performance, the code reviewer will explain
it), be sure to attach a benchmark data here.

Delete section if not applicable.
-->