fix(go): Adjust serialization of nullable value to align python behavior (#2814)
<!--
**Thanks for contributing to Apache Fory™.**

**If this is your first time opening a PR on fory, you can refer to
[CONTRIBUTING.md](https://github.com/apache/fory/blob/main/CONTRIBUTING.md).**

Contribution Checklist

- The **Apache Fory™** community has requirements on the naming of pr
titles. You can also find instructions in
[CONTRIBUTING.md](https://github.com/apache/fory/blob/main/CONTRIBUTING.md).

- Apache Fory™ has a strong focus on performance. If the PR you submit
will have an impact on performance, please benchmark it first and
provide the benchmark result here.
-->

## Why?

<!-- Describe the purpose of this PR. -->

This pr (#2766) introduce new serialization behavior of optional value
in python

This pr (#2793) introduce new hash compute algorithms.

We need to align those behavior in go.

## What does this PR do?

<!-- Describe the details of this PR. -->

- In go, nullable is same as 'referencable', so I reuse the referencable
flag to indicate if a field is nullable. Also, for the string field, we
always treat it as a nullable type but never write its reference.
- rewrite the hash compute function(computeStructHash), now this conform
to the new algorithm in the spec.
- Fix a bug that will cause 'interface' type field to be non-nullable,
which is not expected.

## Related issues

<!--
Is there any related issue? If this PR closes them you say say
fix/closes:

- #xxxx0
- #xxxx1
- Fixes #xxxx2
-->

## Does this PR introduce any user-facing change?

<!--
If any user-facing interface changes, please [open an
issue](https://github.com/apache/fory/issues/new/choose) describing the
need to do so and update the document if necessary.

Delete section if not applicable.
-->

- [x] Does this PR introduce any public API change? no
- [x] Does this PR introduce any binary protocol compatibility change?
no

## Benchmark

<!--
When the PR has an impact on performance (if you don't know whether the
PR will have an impact on performance, you can submit the PR first, and
if it will have impact on performance, the code reviewer will explain
it), be sure to attach a benchmark data here.

Delete section if not applicable.
-->

---------

Co-authored-by: Pan Li <1162953505@qq.com>