feat(rust): Support rust tagged union enum (#2855)
## Why?
Enum with poayload is the norm in rust and is extensively used.


## What does this PR do?
Modified the `Enum::read_data()`, `Enum::write_data()`, and the code
generation for `Enum::default()` to support tagged union enums.

## Related issues
#2851

## Does this PR introduce any user-facing change?
If the user is in `Rust to Rust` protocol instead of `Xlang` protocol,
the values of the tagged union enum can be transmitted correctly.
Otherwise, in `Xlang` protocol, although the corresponding enum type can
be passed, the `internal value` will be the initial value. Because at
present, the `xlang protocol` only serializes the `ordinal value` of
enums.
For example:

```rust
#[test]
fn test_enum(){
    #[derive(ForyObject, Debug, PartialEq)]
    enum Color {
        Red(i8),
        White(String)
    }
    let color = Color::Red(10);
    
    // rust protocol
    let mut fory = Fory::default().xlang(false);
    fory.register::<Color>(1000).unwrap();
    let bin = fory.serialize(&color).unwrap();
    let new_color = fory.deserialize::<Color>(&bin).unwrap();
    assert_eq!(color, new_color);
    if let Color::Red(value) = new_color {
        assert_eq!(value, 10);
    } else {
        unreachable!()
    }
    // xlang protocol
    let mut fory2 = Fory::default().xlang(true);
    fory2.register::<Color>(1000).unwrap();
    let bin = fory2.serialize(&color).unwrap();
    let new_color = fory2.deserialize::<Color>(&bin).unwrap();
    if let Color::Red(value) = new_color {
        assert_eq!(value, i8::default());
    } else {
        unreachable!()
    }
}
```