feat(rust): support rust tuple serialization (#2858)
## What does this PR do?

support rust tuple serialization:
- Compatible mode 
- Consistent mode

Protocol: 
- tuple are taken as LIST for serialization for xlang or compatible mode
- write one by one directly for rust namtive no-compatible mode with no
ref/type meta.
  - For option/pointer/trait, wirte ref meta, type meta then data
  - for others, only write data

Core changes: 
- Use a macro to generate generic tuple serializer for tuple1~20. 
- And update ForyObject macro to support tuple
- Make skip.rs support skip `list/set/map` whose generics is not the
declared generics
- Refactor  skip.rs  to make it more readable and faster

## Related issues
Closes #2863 
Closes #2862

## Does this PR introduce any user-facing change?

<!--
If any user-facing interface changes, please [open an
issue](https://github.com/apache/fory/issues/new/choose) describing the
need to do so and update the document if necessary.

Delete section if not applicable.
-->

- [ ] Does this PR introduce any public API change?
- [ ] Does this PR introduce any binary protocol compatibility change?

## Benchmark

<!--
When the PR has an impact on performance (if you don't know whether the
PR will have an impact on performance, you can submit the PR first, and
if it will have impact on performance, the code reviewer will explain
it), be sure to attach a benchmark data here.

Delete section if not applicable.
-->