feat(c++):  implement xlang serialization for c++ (#2925)
## Why?

#2908 implemented c++ serialization framework, but don't fully implement
xlang spec, and don't align with java/rust.

## What does this PR do?

- implement xlang serialization for c++ 
- align with  java/rust type system
- implement meta string encoding for c++
- add write_type_info/read_type_info
- align collection/map serialization with java/rust
- add xlang serialization tests

## Related issues

#2906

Closes #2909
Closes #1543

## Does this PR introduce any user-facing change?

<!--
If any user-facing interface changes, please [open an
issue](https://github.com/apache/fory/issues/new/choose) describing the
need to do so and update the document if necessary.

Delete section if not applicable.
-->

- [ ] Does this PR introduce any public API change?
- [ ] Does this PR introduce any binary protocol compatibility change?

## Benchmark

<!--
When the PR has an impact on performance (if you don't know whether the
PR will have an impact on performance, you can submit the PR first, and
if it will have impact on performance, the code reviewer will explain
it), be sure to attach a benchmark data here.

Delete section if not applicable.
-->