perf(c++): directly error set instead of result to reduce cost on hotpath (#2959)
## What does this PR do?

directly error set instead of result to reduce cost on hotpath for
buffer read, so we can remove `Result` cost.
## Related issues

#2958 

## Does this PR introduce any user-facing change?

<!--
If any user-facing interface changes, please [open an
issue](https://github.com/apache/fory/issues/new/choose) describing the
need to do so and update the document if necessary.

Delete section if not applicable.
-->

- [ ] Does this PR introduce any public API change?
- [ ] Does this PR introduce any binary protocol compatibility change?

## Benchmark

Before this PR:

| Datatype | Operation | Fory (ns) | Protobuf (ns) | Faster |
|----------|-----------|-----------|---------------|--------|
| Struct | Serialize | 10.7 | 19.8 | Fory (1.9x) |
| Struct | Deserialize | 53.7 | 16.2 | Protobuf (3.3x) |

With this PR:

| Datatype | Operation | Fory (ns) | Protobuf (ns) | Faster |
|----------|-----------|-----------|---------------|--------|
| Struct | Serialize | 10.3 | 19.8 | Fory (1.9x) |
| Struct | Deserialize | 33.6 | 16.5 | Protobuf (2.0x) |